+++ src/wfg/ltv_econ/configs/EconomyConfigLoader.java
@@ -38,8 +38,10 @@ public class EconomyConfigLoader {
         EconomyConfig.MULTI_THREADING = root.getBoolean("MULTI_THREADING");
         EconomyConfig.STARTING_CREDITS_FOR_MARKET = root.getInt("STARTING_CREDITS_FOR_MARKET");
         EconomyConfig.SELF_SUFFICIENCY_REWARD_STRENGTH = root.getDouble("SELF_SUFFICIENCY_REWARD_STRENGTH");
+        EconomyConfig.LOCAL_WORKER_COST_MULT = root.getDouble("LOCAL_WORKER_COST_MULT");
         EconomyConfig.ECON_DEFICIT_COST = root.getDouble("ECON_DEFICIT_COST");
         EconomyConfig.PRODUCTION_BUFFER = 1f + root.getDouble("PRODUCTION_BUFFER");
+        EconomyConfig.LOCAL_PROD_BUFFER = 1f + root.getDouble("LOCAL_PROD_BUFFER");
         EconomyConfig.DAYS_TO_COVER = root.getInt("DAYS_TO_COVER");
         EconomyConfig.DAYS_TO_COVER_PER_IMPORT = root.getInt("DAYS_TO_COVER_PER_IMPORT");
         EconomyConfig.FACTION_EXCHANGE_MULT = (float) root.getDouble("FACTION_EXCHANGE_MULT");
@@ -84,7 +86,9 @@ public class EconomyConfigLoader {
             EconomyConfig.AUTO_TRANSFER_PROFIT_LIMIT = LunaSettings.getDouble(LTV_ECON, "economy_autoTransferLimit").floatValue();
             EconomyConfig.ECON_DEFICIT_COST = LunaSettings.getDouble(LTV_ECON, "economy_deficitCost");
             EconomyConfig.SELF_SUFFICIENCY_REWARD_STRENGTH = LunaSettings.getDouble(LTV_ECON, "economy_selfSufficiencyStrength");
+            EconomyConfig.LOCAL_WORKER_COST_MULT = LunaSettings.getDouble(LTV_ECON, "economy_localWorkerCost");
             EconomyConfig.PRODUCTION_BUFFER = 1f + LunaSettings.getDouble(LTV_ECON, "economy_prodBuffer");
+            EconomyConfig.LOCAL_PROD_BUFFER = 1f + LunaSettings.getDouble(LTV_ECON, "economy_localProdBuffer");
             EconomyConfig.DAYS_TO_COVER = LunaSettings.getInt(LTV_ECON, "economy_daysToCover");
             EconomyConfig.DAYS_TO_COVER_PER_IMPORT = LunaSettings.getInt(LTV_ECON, "economy_toCoverPerImport");
             EconomyConfig.FACTION_EXCHANGE_MULT = LunaSettings.getDouble(LTV_ECON, "economy_factionExchangeMult").floatValue();
@@ -130,9 +134,19 @@ public class EconomyConfigLoader {
         public static double SELF_SUFFICIENCY_REWARD_STRENGTH;
 
         /**
+         * The cost multiplier for workers used to satisfy local demand.
+         */
+        public static double LOCAL_WORKER_COST_MULT;
+
+        /**
          * Applied to the demand vector of worker-independent industries.
          */
         public static double PRODUCTION_BUFFER;
+
+        /**
+         * Applied to the ceiling of local production coefficient.
+         */
+        public static double LOCAL_PROD_BUFFER;
 
         /**
          * each market aims to have <code>x</code> days worth of stockpiles.